Handle std::bad_alloc and invalid formats on RAW file loading
Multi-layer defense against crashes when opening RAW files with invalid
parameters, corrupted headers, or filenames missing resolution info:

1. Resolution sanity boundary: Add Size::MAX_DIMENSION (16384) so that
   width/height validation rejects impossibly large values early,
   preventing integer overflow in memory allocation. FrameHandler
   spinbox maximum updated accordingly.

2. Boundary & safety checks:
   - FileSource::readBytes: reject nrBytes <= 0
   - playlistItemRawFile::loadRawData: guard nrBytes <= 0 and
     fileStartPos < 0
   - videoHandlerRGB/YUV: add isFormatValid() checks before
     loadFrameForCaching, getPixelValue, drawPixelValues,
     calculateDifference
   - videoHandlerYUV::getFormatAsString: guard against empty optional

3. Exception safety: wrap memory-heavy operations (readBytes,
   convertYUVToImage, convertRGBToImage, setFormatFromCorrelation) in
   try-catch for std::bad_alloc. Log via qWarning and exit gracefully
   without updating cache indices. main() wraps app creation to catch
   uncaught bad_alloc.

4. SSE integer overflow fix: cast delta components to int64_t before
   squaring to prevent overflow; change SSE accumulators to uint64_t.
   convertDeltaToRenderValue uses int64_t math with 64-bit clip bounds.

5. User-facing warning in getInfo() when format could not be determined.

PR: IENT/YUView#662

+++ YUViewLib/src/video/rgb/ConversionDifferenceRGB.h
@@ -65,10 +65,10 @@ class SSE
 public:
   void addSample(const rgba_t &delta)
   {
-    this->r += delta.r * delta.r;
-    this->g += delta.g * delta.g;
-    this->b += delta.b * delta.b;
-    this->a += delta.a * delta.a;
+    this->r += static_cast<int64_t>(delta.r) * delta.r;
+    this->g += static_cast<int64_t>(delta.g) * delta.g;
+    this->b += static_cast<int64_t>(delta.b) * delta.b;
+    this->a += static_cast<int64_t>(delta.a) * delta.a;
     ++this->nrSamples;
   }
 
@@ -83,11 +83,11 @@ public:
   }
 
 private:
-  int64_t r{};
-  int64_t g{};
-  int64_t b{};
-  int64_t a{};
-  int64_t nrSamples{};
+  uint64_t r{};
+  uint64_t g{};
+  uint64_t b{};
+  uint64_t a{};
+  uint64_t nrSamples{};
 };
 
 std::pair<QImage, MSE> calculateDifferenceAndMSE(const InputFrameParameters &frame1,
